% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_explainer.R
\name{create_tree_explainer}
\alias{create_tree_explainer}
\title{Create a QSHAPR Tree Explainer}
\usage{
create_tree_explainer(tree_model, max_depth = NULL, base_score = NULL, ...)
}
\arguments{
\item{tree_model}{A tree model object}
}
\value{
A TreeExplainer object
}
\description{
Add this later
}
