% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_summary.R
\name{simple_tree}
\alias{simple_tree}
\title{Simple Tree Structure}
\usage{
simple_tree(
  children_left,
  children_right,
  feature,
  threshold,
  max_depth,
  n_node_samples,
  value,
  node_count
)
}
\description{
Simple Tree Structure
}
