% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_summary.R
\name{tree_summary}
\alias{tree_summary}
\title{Tree Summary Structure}
\usage{
tree_summary(
  children_left,
  children_right,
  feature,
  feature_uniq,
  threshold,
  max_depth,
  sample_weight,
  init_prediction,
  node_count
)
}
\description{
Tree Summary Structure
}
